+++ src/app/(root)/local-food/_components/FixedButtons.tsx
@@ -1,4 +1,4 @@
-import PayButton from '@/components/common/PayButton';
+import PayButton from '@/components/ui/PayButton';
 import { Tables } from '@/types/supabase';
 import { AddCartButton } from './AddCartButton';
